move files to internal folder
