+++ internal/api/api_suite_test.go
@@ -6,7 +6,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	"github.com/env0/terraform-provider-env0/api"
+	"github.com/env0/terraform-provider-env0/internal/api"
 )
 
 var apiClient *api.ApiClient
